greenhouse, rule chain (43)

+++ src/main/resources/files/Greenhouse/finish_soil_acidity.js
@@ -1,7 +1,7 @@
 var newMsg = {
     'ts' : parseInt(metadata.ts),
     'values' : {
-        'acidity' : parseInt(metadata.values_acidity)
+        'acidity' : parseFloat(metadata.values_acidity)
     }
 };
 return {msg: newMsg, metadata: metadata, msgType: msgType};
